update file Router+Method+PopNavigation (Router) New name - Router+Method+UpdateUIViewController
