+++ Sources/Router/Router+Method+UpdateUIViewController.swift
@@ -1,0 +1,20 @@
+/*
+
+Project: SwiftUIRouter
+File: Router+Method+UpdateUIViewController.swift
+Created by: Egor Boyko
+Date: 15.01.2022
+Last Fix: 02.06.2023
+
+Status: #Complete | #Not decorated
+
+*/
+
+import SwiftUI
+
+
+extension Router {
+    public func updateUIViewController(
+        _ pageViewController: UINavigationController,
+        context: Context) {}
+}
